Remove SpheriCart; extract AtomicOrbitals to lib/
