+++ Project.toml
@@ -1,7 +1,7 @@
 name = "Polynomials4ML"
 uuid = "03c4bcba-a943-47e9-bfa1-b1661fc2974f"
 authors = ["Christoph Ortner <christophortner@gmail.com> and contributors"]
-version = "0.5.10"
+version = "0.6.0"
 
 [deps]
 ACEbase = "14bae519-eb20-449c-a949-9c58ed33163e"
@@ -21,13 +21,12 @@ QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
-SpheriCart = "5caf2b29-02d9-47a3-9434-5931c85ba645"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 WithAlloc = "fb1aa66a-603c-4c1d-9bc4-66947c7b08dd"
 
 [compat]
-ACEbase = "0.4.5"
+ACEbase = "0.4.7"
 BenchmarkTools = "1"
 Bumper = "0.7.0"
 ChainRulesCore = "1"
@@ -44,7 +43,6 @@ QuadGK = "2"
 Random = "1.10"
 SparseArrays = "1.10"
 SpecialFunctions = "2.2"
-SpheriCart = "0.2.3"
 StaticArrays = "1.5"
 Test = "1.10"
 WithAlloc = "0.1.0"
