+++ lib/AtomicOrbitals/Project.toml
@@ -1,0 +1,32 @@
+name = "AtomicOrbitals"
+uuid = "326a273f-78e7-476d-aac6-c34ea263ca36"
+authors = ["Christoph Ortner <christophortner@gmail.com> and contributors"]
+version = "0.1.0"
+
+[deps]
+ACEbase = "14bae519-eb20-449c-a949-9c58ed33163e"
+Bumper = "8ce10254-0962-460f-a3d8-1f77fea1446e"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+Polynomials4ML = "03c4bcba-a943-47e9-bfa1-b1661fc2974f"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SpheriCart = "5caf2b29-02d9-47a3-9434-5931c85ba645"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+WithAlloc = "fb1aa66a-603c-4c1d-9bc4-66947c7b08dd"
+
+[compat]
+ACEbase = "0.4.7"
+Bumper = "0.7.0"
+LinearAlgebra = "1.10"
+Polynomials4ML = "0.6"
+Random = "1.10"
+SpheriCart = "0.2.4"
+StaticArrays = "1.5"
+WithAlloc = "0.1.0"
+julia = "1.10"
+
+[extras]
+LuxCore = "bb33d45b-7691-41d6-9220-0943567d0623"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test", "LuxCore"]
