+++ lib/AtomicOrbitals/test/test_atorbrad.jl
@@ -3,11 +3,13 @@ using StaticArrays
 using LinearAlgebra
 using LuxCore
 import Polynomials4ML as P4ML
+import AtomicOrbitals
+import SpheriCart, ACEbase   # ensure the SpheriCart ACEbase extension is active
 using Polynomials4ML: evaluate, evaluate_ed
 
 
 @testset "GaussianBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = P4ML._rand_gaussian_basis()
+    basis = AtomicOrbitals._rand_gaussian_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
@@ -15,7 +17,7 @@ using Polynomials4ML: evaluate, evaluate_ed
 end
 
 @testset "SlaterBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = P4ML._rand_slater_basis()
+    basis = AtomicOrbitals._rand_slater_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
@@ -23,12 +25,9 @@ end
 end
 
 @testset "STOBasis + HyperDual matches evaluate/evaluate_ed" begin
-    basis = P4ML._rand_sto_basis()
+    basis = AtomicOrbitals._rand_sto_basis()
     P4ML.Testing.test_hyperdual_consistency(basis)
     P4ML.Testing.test_evaluate_xx(basis)
     P4ML.Testing.test_chainrules(basis)
     P4ML.Testing.test_withalloc(basis; allowed_allocs = 0, single=false)
 end
-
-
-
